# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/users": {
      "post": {
        "operationId": "UsersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User berhasil dibuat"
          },
          "409": {
            "description": "Email sudah terdaftar"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Buat akun OPERATOR atau ENTRY baru (OWNER)",
        "tags": [
          "Users"
        ]
      },
      "get": {
        "operationId": "UsersController_findAll",
        "parameters": [
          {
            "name": "role",
            "required": false,
            "in": "query",
            "description": "Filter berdasarkan role pengguna",
            "schema": {
              "type": "string",
              "enum": [
                "OPERATOR",
                "ENTRY",
                "OWNER"
              ]
            }
          },
          {
            "name": "is_active",
            "required": false,
            "in": "query",
            "description": "Filter berdasarkan status aktif. true = aktif, false = nonaktif",
            "schema": {
              "example": true,
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Daftar user berhasil diambil"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lihat daftar semua user di merchant (OWNER)",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users/{id}": {
      "patch": {
        "operationId": "UsersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users/{id}/status": {
      "patch": {
        "operationId": "UsersController_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Status user berhasil diubah"
          },
          "403": {
            "description": "User bukan milik merchant ini"
          },
          "404": {
            "description": "User tidak ditemukan"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Aktifkan atau nonaktifkan akun user (OWNER)",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users/{id}/change-password": {
      "post": {
        "operationId": "UsersController_changePassword",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Users"
        ]
      }
    },
...
    "/api/v1/transactions": {
      "get": {
        "operationId": "TransactionsController_findAll",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "PENDING",
                "CONFIRMED"
              ]
            }
          },
          {
            "name": "sync_status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "PENDING_SYNC",
                "SYNCING",
                "SYNCED",
                "SYNC_FAILED",
                "SYNC_CONFLICT"
              ]
            }
          },
          {
            "name": "id_device",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "start_date",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "end_date",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Cursor for pagination (id_transaction of the last item from previous page)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Transactions"
        ]
      }
    },
    "/api/v1/transactions/{id}": {
      "get": {
        "operationId": "TransactionsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Transactions"
        ]
      }
    },
    "/api/v1/transactions/{id}/void": {
      "post": {
        "operationId": "TransactionsController_void",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VoidTransactionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Transactions"
        ]
      }
    },
    "/api/v1/transactions/{id}/conflict-resolution": {
      "post": {
        "operationId": "TransactionsController_resolveConflict",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveConflictDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Transactions"
        ]
      }
    },
    "/api/v1/transactions/{id}/resolve": {
      "post": {
        "operationId": "TransactionsController_resolveConflictCompatibility",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveConflictDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Transactions"
        ]
      }
    },
    "/api/v1/transactions/{id}/corrections": {
      "post": {
        "operationId": "TransactionsController_correct",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CorrectTransactionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Transactions"
        ]
      }
    },
    "/api/v1/transactions/{id}/correct": {
      "post": {
        "operationId": "TransactionsController_correctCompatibility",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CorrectTransactionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Transactions"
        ]
      }
    },
...
  },
  "info": {
    "title": "K-POS API",
    "description": "Canonical API for the K-POS multi-app offline-first platform",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "full_name": {
            "type": "string",
            "example": "Andi Kasir"
          },
          "email": {
            "type": "string",
            "example": "andi@toko.com"
          },
          "password": {
            "type": "string",
            "example": "password123",
            "minLength": 8
          },
          "role": {
            "type": "string",
            "enum": [
              "OPERATOR",
              "ENTRY"
            ],
            "example": "OPERATOR",
            "description": "Role yang diizinkan: OPERATOR (kasir) atau ENTRY (staf input). OWNER tidak bisa dibuat via endpoint ini."
          }
        },
        "required": [
          "full_name",
          "email",
          "password",
          "role"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "full_name": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "ENTRY",
              "OPERATOR"
            ]
          }
        }
      },
      "UpdateUserStatusDto": {
        "type": "object",
        "properties": {
          "is_active": {
            "type": "boolean",
            "example": false,
            "description": "true = aktifkan akun, false = nonaktifkan akun"
          }
        },
        "required": [
          "is_active"
        ]
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "new_password": {
            "type": "string",
            "minLength": 8
          }
        }
      },
...
      "VoidTransactionDto": {
        "type": "object",
        "properties": {
          "void_reason": {
            "type": "string",
            "description": "Reason for voiding the transaction",
            "example": "Customer left their wallet",
            "minLength": 5
          }
        },
        "required": [
          "void_reason"
        ]
      },
      "ResolveConflictDto": {
        "type": "object",
        "properties": {
          "action": {
            "type": "string",
            "enum": [
              "CONFIRM",
              "VOID"
            ],
            "description": "CONFIRM = Paksa konfirmasi transaksi (stok akan dipotong meski negatif). VOID = Batalkan transaksi yang konflik.",
            "example": "CONFIRM"
          },
          "notes": {
            "type": "string",
            "description": "Catatan rekonsiliasi untuk audit trail",
            "example": "Barang sudah diberikan ke pelanggan, stok disetujui meski negatif",
            "minLength": 5
          }
        },
        "required": [
          "action"
        ]
      },
      "CorrectItemDto": {
        "type": "object",
        "properties": {
          "id_product": {
            "type": "string",
            "example": "clxxxxx",
            "description": "ID produk"
          },
          "quantity": {
            "type": "number",
            "example": 2,
            "description": "Jumlah unit produk"
          },
          "unit_price": {
            "type": "number",
            "example": 15000,
            "description": "Harga per unit (dalam rupiah)"
          },
          "subtotal": {
            "type": "number",
            "example": 30000,
            "description": "Subtotal baris ini (quantity * unit_price)"
          }
        },
        "required": [
          "id_product",
          "quantity",
          "unit_price",
          "subtotal"
        ]
      },
      "CorrectTransactionDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "description": "Alasan koreksi yang wajib diisi untuk audit trail",
            "example": "Kasir salah input kuantitas produk A",
            "minLength": 10
          },
          "items": {
            "description": "Daftar item yang benar (akan menggantikan item lama)",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CorrectItemDto"
            }
          },
          "subtotal": {
            "type": "number",
            "example": 45000,
            "description": "Subtotal transaksi koreksi"
          },
          "total": {
            "type": "number",
            "example": 45000,
            "description": "Total transaksi koreksi"
          },
          "notes": {
            "type": "string",
            "example": "Transfer BCA",
            "description": "Catatan tambahan koreksi"
          }
        },
        "required": [
          "reason",
          "items",
          "subtotal",
          "total"
        ]
      },
...
    }
  }
}